refactor(worker): narrow worker module exports
