+++ src/worker/worker-rpc-clients.ts
@@ -1,15 +1,3 @@
-export {
-  type WorkerInferenceHandlers,
-  WorkerInferenceProxyClient,
-  WorkerInferenceProxyError,
-} from "./worker-rpc-inference-client.js";
-export {
-  WorkerLiveEventClient,
-  type WorkerLiveEventClientOptions,
-  WorkerLiveEventError,
-} from "./worker-rpc-live-event-client.js";
-export {
-  WorkerTranscriptCommitClient,
-  type WorkerTranscriptCommitClientOptions,
-  WorkerTranscriptCommitError,
-} from "./worker-rpc-transcript-client.js";
+export { WorkerInferenceProxyClient } from "./worker-rpc-inference-client.js";
+export { WorkerLiveEventClient } from "./worker-rpc-live-event-client.js";
+export { WorkerTranscriptCommitClient } from "./worker-rpc-transcript-client.js";
