+++ src/worker/worker.runtime.ts
@@ -1,13 +1,7 @@
 import { chmod, mkdtemp, realpath, rm, stat } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import path from "node:path";
-import type { Readable, Writable } from "node:stream";
-import { WORKER_PROTOCOL_MAX_INFERENCE_PAYLOAD_BYTES } from "../../packages/gateway-protocol/src/schema/worker-inference.js";
-import {
-  buildWorkerConnectParams,
-  parseWorkerLaunchDescriptor,
-  type WorkerLaunchDescriptor,
-} from "./launch-descriptor.js";
+import { buildWorkerConnectParams, type WorkerLaunchDescriptor } from "./launch-descriptor.js";
 import { createWorkerConnection, type WorkerConnectionState } from "./worker-connection.js";
 import {
   WorkerInferenceProxyClient,
@@ -15,50 +9,14 @@ import {
   WorkerTranscriptCommitClient,
 } from "./worker-rpc-clients.js";
 
-export type WorkerRuntimeResult =
+type WorkerRuntimeResult =
   | { status: "completed"; transcriptLeafId: string | null; transcriptNextSeq: number }
   | { status: "fenced"; reason: "credential-replaced" | "owner-epoch-mismatch" };
-
-export type RunWorkerCommandOptions = {
-  input: Readable;
-  output: Writable;
-};
 
 const WORKER_REMOTE_CANCEL_GRACE_MS = 1_000;
 
 function toError(value: unknown, fallback: string): Error {
   return value instanceof Error ? value : new Error(fallback, { cause: value });
-}
-
-async function readLaunchDescriptor(input: Readable): Promise<WorkerLaunchDescriptor> {
-  const chunks: Buffer[] = [];
-  let byteLength = 0;
-  for await (const rawChunk of input as AsyncIterable<unknown>) {
-    const chunk =
-      typeof rawChunk === "string"
-        ? Buffer.from(rawChunk)
-        : rawChunk instanceof Uint8Array
-          ? Buffer.from(rawChunk)
-          : undefined;
-    if (!chunk) {
-      throw new Error("worker launch descriptor input must be bytes");
-    }
-    byteLength += chunk.byteLength;
-    if (byteLength > WORKER_PROTOCOL_MAX_INFERENCE_PAYLOAD_BYTES) {
-      throw new Error("worker launch descriptor exceeds the protocol payload limit");
-    }
-    chunks.push(chunk);
-  }
-  if (byteLength === 0) {
-    throw new Error("worker launch descriptor is required on stdin");
-  }
-  let decoded: unknown;
-  try {
-    decoded = JSON.parse(Buffer.concat(chunks).toString("utf8")) as unknown;
-  } catch (error) {
-    throw new Error("worker launch descriptor is not valid JSON", { cause: error });
-  }
-  return parseWorkerLaunchDescriptor(decoded);
 }
 
 function fencedResult(state: WorkerConnectionState): WorkerRuntimeResult | undefined {
@@ -224,21 +182,5 @@ export async function runWorkerDescriptor(
       process.env.OPENCLAW_CONFIG_PATH = previousConfigPath;
     }
     await rm(stateDir, { recursive: true, force: true });
-  }
-}
-
-export async function runWorkerCommand(options: RunWorkerCommandOptions): Promise<void> {
-  const descriptor = await readLaunchDescriptor(options.input);
-  const abortController = new AbortController();
-  const stop = () => abortController.abort(new Error("worker interrupted"));
-  process.once("SIGINT", stop);
-  process.once("SIGTERM", stop);
-  try {
-    const result = await runWorkerDescriptor(descriptor, { signal: abortController.signal });
-    const encoded = `${JSON.stringify(result)}\n`;
-    options.output.write(encoded);
-  } finally {
-    process.off("SIGINT", stop);
-    process.off("SIGTERM", stop);
   }
 }
